requirements.txt
aiofiles==23.2.1
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
Django==5.0.4
django-elasticsearch-dsl==8.0
elastic-transport==8.13.0
elasticsearch==8.13.0
elasticsearch-dsl==8.13.1
etelemetry==0.3.1
filelock==3.13.1
fitz==0.0.1.dev2
fonttools==4.51.0
frontend==0.0.3
fsspec==2024.3.1
GoogleTran==1.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.23.0
hyperframe==5.2.0
idna==2.10
intel-openmp==2021.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
langdetect==1.0.9
looseversion==1.3.0
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.8.4
mkl==2021.4.0
mod_wsgi==5.0.0
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pillow==10.3.0
platformdirs==4.2.0
prov==2.0.0
pydot==2.0.0
pyotp==2.9.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypng==0.20220715.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytils==0.4.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
qrcode==7.4.2
rdflib==7.0.0
regex==2024.4.28
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.3
scipy==1.13.0
sentencepiece==0.2.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
starlette==0.37.2
sympy==1.12
tbb==2021.12.0
tokenizers==0.19.1
tools==0.1.9
torch==2.3.0
tqdm==4.66.4
traits==6.3.2
transformers==4.40.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.0
